Refactor ProviderAPIs object to better design pattern
Utilize a shared state design pattern (Borg) instead of singleton.
This is a significantly better design pattern as it allows the
ProviderAPIRegistry to be non-private and instantiated as needed
but maintain exactly the same state across all the objects (data
is stored on the class definition).

A __setattr__ method has been implemented to avoid accidental
changes or data creep to end up on the Registry itself.

Change-Id: Ic7c84af014bb20158d7c23340da9d8cdae3908c8