Remove unclear wording in parameters
This was brought up in a previous change[0], but these parameters
are worded to detail that they each contain varius attributes. This
works when listing them in the yaml file, but the meaning is lost
when these parameters are rendered into tables and becomes slightly
confusing.

This change removes the various wording of "contains the following"
from the various object parameters.

[0] https://review.openstack.org/#/c/569741/3

Change-Id: Ibced08ba0c2e5f8cbadbacc675dc0cf6b745ecdc