Add new "How Can I Help?" contributor guide
We see a lot of very small patches from new contributors to do things
like fix typos, correct grammar, switch to HTTPS URLs, and similar
trivial changes. Many of these changes are technically correct, but the
value they provide to the project is very low. Assuming the best of
intentions, the obvious conclusion is that these contributors very much
want to help, but don't know where to start and haven't been given
proper guidance.

This patch introduces a new guide that lists some ongoing tasks that we
always want help with. These should be tasks that someone with no
historical context in the keystone project should be able to pick up,
but by starting a few steps above typo fixes will be able to learn
enough to eventually make more significant contributions. When typo fix
patches come in, we can use this guide to gently encourage new people to
make more valuable contributions.

Change-Id: I2a003bbef3ffbde1818fddcb840905eac6f0618c