Refactor - remove extra for loop
This change refactors the logic for obtaining methods from an
integer. Instead of creating a list of integers to later loop
through in order to create a list of methods, skip ahead to
creating a list of methods since we already have the int key and
can pull the method from the map.

Change-Id: I4fc5b23b8f82d0f475a0ca787f2df10af1ed7238