Refactor _set_domain_id_and_mapping functions
This change refactors both functions listed for single refs and
a list of refs for _set_domain_id_and_mapping to do separate checks
for conditions such as _is_domain_aware and _is_mapping_needed. This
is due to how it is currently handled, it uses the same logic between
both functions, which provides more overhead for a list than a single
ref.

Change-Id: I325627f511fe7d52a191e64eada030e84d756d7e