Expose random uuid bug in cadf notifications
When enabling CADF notifications and clearing the notification_opt_out
setting[0] (which cause keystone to be more chatty with notifications)
in order to audit identity.authenticate events, keystone emits a
notification for the identity.authentication event where the
initiator's ID is a random UUID that doesn't match up to the user
who authenticated.

This was noticed when users would log into horizon and from the audit
notification, it could not be determined which user was logging in.

This appears to be true for only the authentication event, other
events such as project CRUD appear to work as intended.

[0] https://github.com/openstack/keystone/blob/master/keystone/conf/default.py#L221

Change-Id: I8363f265faacb9ad5783b12d4a64b570b9af46f0
Related-Bug: #1780503