Allow class-level definition of API URL Prefix
Add the ability to define API URL Prefix at the class-level. This
will be used only if no ap_url_prefix is passed into __init__.

Flask-restful handles prefixes in the order of `BAE` as used in Keystone:
  * Blueprint
  * API
  * Endpoint (flask-endpoint)

This allows for additional prefixes such as "OS-TRUST" for
legacy/converted extension code that has not been updated to use
non-"OS" prefixed pathes.

Change-Id: I7acf2f16a25efc6a801e6c87c1aac436bb8e5453
Partial-Bug: #1776504