Allow for 'extension' rel in json home
Allow providing an alternative REL function (e.g. for extensions)
in JSON Home in flask. The alternative functions are expected to
be functools.partial.

The extension-mode should only ever be used in the case of a keystone
API that was an extension (legacy style) and has not been converted
over to a new URI path; typically extensions start with an OS- prefix
in the URL.

All new APIs should not be using the 'extension' style.

Change-Id: I3970aef12f3b7c8e84fa80dfea8fc6eda63fe77d
Partial-Bug: #1776504