Pass path into full_url and base_url
This change mirrors what was done prior to flask and leans on the
base_url and full_url functions to append '/v3' and the path or
collection_name. This will allow for changes to wrap_member and
wrap_collection to properly handle local-only wrapping (no need
to cross-api call wrap because a "role" is returned from the "trust"
api).

Change-Id: I5376d8abc581994f28da64f3b578a04b00e68c4d
Partial-Bug: #1776504