Make policy file support in fixture optional
The ksfixtures.Policy object used to accept two arguments. One was
the path of the policy file and the other was the test fixture. Since
policy is now kept in code and the default policy file has been
removed from keystone source, we can simplify the fixture to
optionally deal with a file path, instead of passing a non-existant
file to it all the time.

Change-Id: I9d8c4cbf963099fe73e39dbf46e03f66f9a79f43