Loosen the assertion for logging scope type warnings
We have tests to make sure we raise a 403 Forbidden in the event we
catch an exception from oslo.policy due to InvalidScope. An assertion
in these tests made sure we logged a warning by checking a mock was
called exactly once.

This is going to fail when we start deprecating policies in Stein in
favor of ones that enable system scope checking. Instead of checking
that assertion is called only once, let's just make sure it is called
at least once with the arguments we expect. Otherwise we're going to
catch false positives with the assertion because deprecated policies
are also logged for operators.

Change-Id: I5da718d5b31e90f1aef7d5b457b3c772eb57d1fb