Ensure view args is in policy dict
The policy_dict (in enforcement) was not populating the view args
in a similar manner to the old style @protected decorator. This
change ensures that we mirror the old behavior (required for
proper use of v3cloud policy).

Conflicts:
    keystone/tests/unit/common/test_rbac_enforcer.py

Change-Id: Ida9009a95a874be9cc60c3152d4e3225726562eb
Partial-Bug: #1776504
Closes-Bug: #1792047
(cherry picked from commit 4975b79e8174587f7639347939cf679460d4896b)