LDAP attribute names non-case-sensitive
keystone was not able to find any users while
the LDAP user name attribute was configured to
"samaccountname", but could find users when
reconfigured to use "sAMAccountName". LDAP is
not supposed to be case-sensitive, so either
should work.

This patch addresses the above problem by making
both the attributes into lower case. Also updated
the ldap_result example supporting python3.

Change-Id: I51813ac41489baed04f3cadbccd748e03025313e
Closes-Bug: #1753585