Comment out un-runnable tests
These tests have not been run in > 2 years. They are commented out
with an updated FIXME to rework once the flask port is done (auth).
It is out of scope of Flask to re-enable long disabled tests.
We do not want to lose the context of the coverage the tests provide
thus we are commenting them out instead of outright deletion.

Change-Id: I0760746dc62b65607ac0e88ee6d03395c9226fe7