Add python3 functional test job
As part of the python3-first goal, this patch ensures we have a voting
functional job that uses python3. The choice to call it py3 rather than
py35 is conscious, as the underlying python3 version is controlled by
the choice of test node which the job inherits from its ancestor jobs.

Change-Id: I97ced2047964055966fd6c2e8acfec29b48bdfe3