Allow domain users to access the GET domain API
This change updates the policy for identity:get_domains to allow
users with role assignments on a domain to fetch the domain. As long
as they call the API with a domain scoped token, they should be
authorized to retrieve the domain. Subsequent patches will do the
same for project users.

Change-Id: I83a9b8af775580d36a1141be55e9c1cc283a75b6
Partial-Bug: 1794864
Partial-Bug: 968696