Allow project users to retrieve domains
This commit adds thorough testing to make sure users who have a role
on a project can use project-scoped tokens to call GET
/v3/domain/{domain_id} for the domain own their project. These users
are not allowed to access domains that they don't have any
authorization via project role assignments.

This ensures the domains API is tested with these cases and makes the
domains API more self-serviceable for users that are not
administrators.

Change-Id: Ifc100a7a235140fbd07cbafe80983d3c2f17a7dc
Closes-Bug: 1794864
Related-Bug: 968696