Avoid using dict.get() in assertions
The method dict.get(key) will return None
in the following two cases:

A. if there is no key in the dictionary.
B. if the real value of the key is just None.

The above two cases will results in passing
the assertion which is not the expected
behaviour of the test.

Change-Id: Ib4a7f2d5beb1e4e9b4d8d9dd90de3f52a5c29e74