Clarify group-mapping example in docs
The mapping example adds mapped users to a group, so it is misleading to
say that it grants them a role directly, which is the purview of
autoprovisioning. This patch clarifies that, and also updates the role
name to be `reader` which is the new default read-only role.

Followup to https://review.openstack.org/605496

Change-Id: I669df0800275c49ed1ec2c3934301313fb7c4dca