Don't mock internal implementation details of oslo
Keystone relies on oslo.messaging to handle the underlying routing
of messaging for keystone notifications and CADF events. Early
testing implementations for notification support in keystone mocked
and internal method of the oslo.messaging implementation to make sure
it was actually being invoked. This falls outside of the realm of
keystone testing because it's testing the internal implementation
details of an external library (which is already tested anyway).

This commit adjusts the tests to mock a public facing method instead
of a private one. This makes the mock safer since it's consuming a
public API just like the rest of the notification framework.

Change-Id: I04b78a94eddda5486337e7555817a25e0a106433