Update notification tests to work with o-m 9.0.0
oslo.messaging 9.0.0 was released with a broken default for
transport_url that caused some of our tests to fail. This commit
updates those tests classes to override transport_url with sane
value, which is the default reintroduced into oslo.messaging in
version 9.0.1.

This just makes it so that we don't have to blacklist version 9.0.0
from keystone's requirements.

Change-Id: I2b43a6c81244c1956e3a943dcbc1cae0e3631eda