Set Default and resource limit as defined schema
Default_limit and Resource_limit should not be set
as negative values or out of range. This patch
addresses the above issue be checking the limit
passed by the user should be minimum and maximum to
the values defined in schema. If not, then raising
the 400 bad request. Also added the test cases for the
same.

Change-Id: I089f3c231a38be1a1f769a35b4904f5c078d5d07
Closes-Bug: #1797876