Refactor flask domain config resources
The domain config flask resource classes all call the same
config_api with the same arguements that are provided if present.

This change removes the original duplication that was present
to handle this, instead replacing it with subclasses for each
API route to be handled with flask.

Also fixed a copy/paste fail for s/POST/PUT

Change-Id: I5a6194fb50ee49fed01bea960507d9b227c9129e