Delete administrator federation guide
The federated-identity guide in the admin section is out of date
relative to the one in the operator section. It is sort of incoherent
and the information within it that is valuable is already present in the
other guide. Delete this guide in favor of just promoting the operator
guide.

Related-bug: #1793374

Change-Id: I8c0190725d0f7297957bbf10fd9d7184e52ff5f3