Provide a Location on HTTP 300
To best adhere to the RFC2616, we should emit a Location header when
issuing an HTTP 300 to point to the preferred version. While we only
have a single version active it is 'v3'. In the future it should
always be the most recent set of CRUD. This location is strictly for
CRUD purposes and will lean on the WWW-Authentication URI to point
to the most correct AUTH uri.

Change-Id: Ibdd53f236a3c51d1aa23eac35dd595e2dae79ba6
closes-bug: #1230927