Remove useless use of :orphan:
The federation docs have **orphan:** in bold printed on the published
pages. The :orphan: directive isn't needed for anything, so remove it to
keep it from showing up in the rendered docs.

The Caching Layer docs use the :orphan: directive as well and don't have
this issue, but also don't seem to need it there, so remove it from
there as well.

Change-Id: Icea1a89e04e4c249ba1d78f95ff5a66b9a297284