Refactor directory creation into a common place
There is a useful method in the fernet_utils.py module for creating
directories. We can reuse this for the jwt provider implementation.

This commit generalizes the logic and moves it into a common place
for fernet_utils to use. A subsequent patch introducing JWT
functionality will use the same utils.

bp json-web-tokens

Change-Id: I0e322a745683c43e8329110a165f54d2e998e0c7