Update more info of vhost file
The federation document talks about the
vhost file to be changed when configuring
shibboleth. As a user it is difficult to
understand where the file is located and
when configuring through devstack no
vhost file of name *keystone.conf* exists.

Thus this patch adds more information regarding
vhost file for more understandability.

Change-Id: Ifb43e6532e8d388bfcfee64d36e43e5120aab4c5