Add mapping tests for system member role
From keystone's perspective, the ``member`` and ``reader`` roles are
effectively the same. This is primarily because the member role is
really meant for project members and project-specific resources, which
doesn't apply to mapping resources.

This commit adds explicit testing to make sure the ``member`` role
is allowed to perform readable and not writable mapping operations.
Subsequent patches will incorporate:

 - system admin functionality
 - testing for domain users
 - testing for project users

 Related-Bug: 1804519
 Related-Bug: 1804521

Change-Id: I8a7ecd37f4db59fb8e10b68b03bbaea543484e6d