Implement system reader role for users
This commit introduces the system reader role to the users API, making
it easier for administrators to delegate subsets of responsibilities
to the API by default. This commit also maintains the ability for
any user to be able to fetch their own user reference.

Subsequent patches will incorporate:

  - system member test coverage
  - system admin functionality
  - domain reader functionality
  - domain member test coverage
  - domain admin functionality
  - project user test coverage

Change-Id: I9c362e515772540dfa93d05781d955009b9a154d
Partial-Bug: 1805406
Partial-Bug: 1748027
Partial-Bug: 968696