Restructure federation guide
Having everything on a single page is nice for ctrl-F-ability but it
makes the flow very confusing. This change breaks the guide into three
logical parts: the introduction, the configuration steps, and the
advanced mapping rules guide. Keeping all the configuration steps within
one page means it can still be searched easily, but removing the prose
of the introduction and breaking out the deep-dive mapping rules guide
reduces clutter and enhances readability.

Partial-bug: #1793374

Change-Id: Id2fd59d62a2675691d545e4cd0404558f00cf481