Move "Public ID Generators" to relevant docs
Currently, the section "Public ID Generators" is a subsection of
"Identity sources" but it reads as very out of place. Looking at the
commit that introduced the section (1a50986e7c), it's clear this was
meant to be part of the domain-specific-config section and was missed in
a reshuffle. This patch puts it back in place.

Change-Id: I2873f104adf6af4da4ba23f8c0d8afb0c1161da3