Consolidate WebSSO guide into SP instructions
The WebSSO documentation does not have a good flow with the rest of the
federation guide. It includes instructions about the remote_id_attribute
which is not WebSSO specific, as well as instructions for configuring
Apache which is partly redundant with the SP-specific instructions. This
change consolidates the guide into the main guide so that it makes sense
with the rest of the document.

Partial-bug: #1793374

Change-Id: I0c8fa537a950090f85b3cb4a4aac6c896f02db89