Add configuration options for JWS provider
This commit only introduces a new group for JWT configuration options
that are specific to JWS. Even though the configuration group is named
after JWT, the configuration options are specific to JWS. If keystone
supports JWE token in the future, we will add new configuration
options for key specific to that implementation as opposed to reusing
JWS keys.

It does not wire anything up to the actual token provider code. That
will come in a later patch.

bp json-web-tokens

Change-Id: I8c9b249138a32f096d5edc84772099e9d780687b