Add documentation for writing specifications
We have a specification template and a keystone-specs repository, but
we don't offer any documentation or guidance on the specification
process we have.

This commit aims to clarify the difference between specifications and
RFE bug reports, and why we use both. It also describes the process
someone should follow if they want to file a feature request against
the project.

Change-Id: I5dbf6f81058de3f2a64a95e4cf34a1279a49c5dd