fix self-service credential APIs
Self-service credential APIs are broken in stable/rocky because
it no longer building the target attributes with the entity
data. Therefore, only "admin" can perform credential optionations.
The lack of self-service capability limits the usefulness of
these APIs. Arguably this is break backward compatibility as
self-service used to work in stable/queens and older releases.
Though we've never built the properly tests to guard this
functionality.

This patch re-enables self-service capability by conveying the
entity data via the target attributes. It also build the proper
tests for it.

Change-Id: Ic7dddc4d2fe7b6c6ae3bf6aed6c4c048743b3eed
Closes-Bug: 1815539