Fix mock for v2 test
In 7eec2c5d we removed the _VERSIONS constant from the api discovery
module. By a quirk of the way the unit tests are ordered, the mock to
patch this constant continued to work if the whole test suite was run,
but it doesn't work if we just run that test alone or the test_versions
suite alone. This patch cleans up the unneeded mock.

Change-Id: If88db5b8f4cb3425fd7a568213aa094fe7800c38